refactor(frontend): corregir violaciones de capa en GrillaPage [US-ADJ-6.5]
FE-ARCH-02: extrae acceso a db/queries en useGrillaQueue.ts — GrillaPage ya
no importa de db/ directamente. FE-DES-01: mueve formatMarca/buildResultadoValue
a utils/marca.ts — GrillaPage y AtletaCard importan desde utils/.

tsc 0 errores · build exitoso · ESLint 0 errores.
