+++ frontend/src/utils/marca.ts
@@ -1,0 +1,20 @@
+function buildRpValue(metros: number, centimetros: string): string {
+  return `${metros}.${centimetros.padEnd(2, '0')}`
+}
+
+export function buildResultadoValue(metros: number, centimetros: string, unidad: string): string {
+  if (unidad === 'Segundos') {
+    return String(metros * 60 + Number(centimetros || '0'))
+  }
+  return buildRpValue(metros, centimetros)
+}
+
+export function formatMarca(value: string, unidad: string): string {
+  if (unidad === 'Segundos') {
+    const totalSeconds = Number(value || '0')
+    const minutos = Math.floor(totalSeconds / 60)
+    const segundos = totalSeconds % 60
+    return `${minutos}:${String(segundos).padStart(2, '0')} min`
+  }
+  return `${value} m`
+}
